+++ arch/arm/src/stm32/stm32_oneshot.c
@@ -259,9 +259,8 @@ int stm32_oneshot_start(struct stm32_oneshot_s *oneshot,
   uint64_t period;
   irqstate_t flags;
 
-  tmrinfo("handler=%p arg=%p, ts=(%lu, %lu)\n",
-         handler, arg, (unsigned long)ts->tv_sec,
-         (unsigned long)ts->tv_nsec);
+  tmrinfo("handler=%p arg=%p, ts=(%jd, %ld)\n",
+         handler, arg, (intmax_t)ts->tv_sec, ts->tv_nsec);
   DEBUGASSERT(oneshot && handler && ts);
   DEBUGASSERT(oneshot->tch);
 
@@ -449,8 +448,8 @@ int stm32_oneshot_cancel(struct stm32_oneshot_s *oneshot,
           ts->tv_nsec = nsec;
         }
 
-      tmrinfo("remaining (%lu, %lu)\n",
-             (unsigned long)ts->tv_sec, (unsigned long)ts->tv_nsec);
+      tmrinfo("remaining (%jd, %ld)\n",
+              (intmax_t)ts->tv_sec, ts->tv_nsec);
     }
 
   return OK;
