+++ arch/sparc/src/bm3803/bm3803_oneshot.c
@@ -206,8 +206,8 @@ int bm3803_oneshot_start(struct bm3803_oneshot_s *oneshot,
   uint64_t period;
   irqstate_t flags;
 
-  tmrinfo("handler=%p arg=%p, ts=(%lu, %lu)\n", handler, arg,
-          (unsigned long)ts->tv_sec, (unsigned long)ts->tv_nsec);
+  tmrinfo("handler=%p arg=%p, ts=(%jd, %ld)\n", handler, arg,
+          (intmax_t)ts->tv_sec, ts->tv_nsec);
   DEBUGASSERT(oneshot && handler && ts);
   DEBUGASSERT(oneshot->tch);
 
@@ -393,8 +393,8 @@ int bm3803_oneshot_cancel(struct bm3803_oneshot_s *oneshot,
           ts->tv_nsec = nsec;
         }
 
-      tmrinfo("remaining (%lu, %lu)\n",
-             (unsigned long)ts->tv_sec, (unsigned long)ts->tv_nsec);
+      tmrinfo("remaining (%jd, %ld)\n",
+              (intmax_t)ts->tv_sec, ts->tv_nsec);
     }
 
   return OK;
