+++ arch/x86_64/src/intel64/intel64_oneshot.c
@@ -290,9 +290,8 @@ int intel64_oneshot_start(struct intel64_oneshot_s *oneshot,
   uint64_t   compare = 0;
   irqstate_t flags;
 
-  tmrinfo("handler=%p arg=%p, ts=(%lu, %lu)\n",
-         handler, arg, (unsigned long)ts->tv_sec,
-         (unsigned long)ts->tv_nsec);
+  tmrinfo("handler=%p arg=%p, ts=(%jd, %ld)\n",
+         handler, arg, (intmax_t)ts->tv_sec, ts->tv_nsec);
 
   DEBUGASSERT(oneshot && handler && ts);
   DEBUGASSERT(oneshot->tch);
@@ -464,8 +463,8 @@ int intel64_oneshot_cancel(struct intel64_oneshot_s *oneshot,
       ts->tv_sec  = sec;
       ts->tv_nsec = nsec;
 
-      tmrinfo("remaining (%lu, %lu)\n",
-             (unsigned long)ts->tv_sec, (unsigned long)ts->tv_nsec);
+      tmrinfo("remaining (%jd, %ld)\n",
+              (intmax_t)ts->tv_sec, ts->tv_nsec);
     }
 
   return OK;
